reorganized code into folders and renamed sanitize_data to run_sanitization
