+++ code/steps/sanitize_urls/classes/infokind.py
