+++ code/steps/sanitize_urls/run_sanitization.py
@@ -2,25 +2,17 @@
 Sanitizes contact info from a database and places the results in another database.
 '''
 import logging
-import sqlalchemy
 import json
-from enum import Enum
 import sys
 import argparse
-from collections import namedtuple
-from typing import List, Dict, Tuple, Union
-from collections import OrderedDict
-import phonenumbers
-from email_validator import validate_email
-import pprint
 
-from pg_sanitization import *
-from sanitize_urls import get_sanitized_urls_for_update
-from sanitize_phone_nums import get_sanitized_phone_nums_for_update
-from sanitize_emails import get_sanitized_emails_for_update
-from helper_methods import positive_int
-from infokind import InfoKind
-from src2dest import Src2Dest
+from sanitization_code.pg_sanitization import *
+from sanitization_code.sanitize_urls import get_sanitized_urls_for_update
+from sanitization_code.sanitize_phone_nums import get_sanitized_phone_nums_for_update
+from sanitization_code.sanitize_emails import get_sanitized_emails_for_update
+from sanitization_code.helper_methods import positive_int
+from classes.infokind import InfoKind
+from classes.src2dest import Src2Dest
 
 this_module = sys.modules[__name__]
 
@@ -91,7 +83,8 @@ def main():
     dest_conn = None
     if args.dest_db:
         dest_db, dest_conn = get_db_conn(args.dest_db)
-        
+    
+    # for each source-dest-mapping, sanitize data in source table/database and place into dest table/database
     for s2d in args.source_dest_mapping:
         logger.info(f'Mapping \n{s2d}')
         
@@ -101,6 +94,7 @@ def main():
         if s2d.kind not in ['url', 'email', 'phone']:
             raise Exception(f"Unknown kind={s2d.kind}")
         
+        # sanitizing data in batches defined by batch-row-size
         for offset in range(0, args.max_rows_to_fetch, args.batch_row_size):
             logger.info(f"Fetching from {offset}")
             print(dict(
@@ -110,6 +104,7 @@ def main():
                     offset=offset
                 ))
 
+            # query data to be sanitized with the specified keys
             results = query_db(
                 src_conn, 
                 table = s2d.source_table, 
@@ -119,12 +114,15 @@ def main():
                 offset = offset
             )   
             
+            # if no results, break loop because we have reached the end of the data table
             if not len(results):
                 break
 
             key_vals = [result[1:] for result in results]
             raw_data = [result[0] for result in results]
 
+            # get sanitized data JSONs that contain the IDs and sanitized urls/phone_nums/emails
+            # these JSONs will be used to update the raw_data in a SQL UPDATE statement
             if s2d.kind == 'url': 
                 sanitized_data = get_sanitized_urls_for_update(raw_data, keys = s2d.key, key_vals = key_vals, source_table = s2d.source_table, source_column = s2d.source_column, infokind = s2d.kind, logger = logger)
             elif s2d.kind == 'phone':
@@ -133,6 +131,7 @@ def main():
                 sanitized_data = get_sanitized_emails_for_update(raw_data, keys = s2d.key, key_vals = key_vals, source_table = s2d.source_table, source_column = s2d.source_column, infokind = s2d.kind, logger = logger)
 
             if not args.write:
+                # if there is not destination table/database specified, then update raw data in the source table/database with sanitized output
                 if not dest_conn:
                     update_src_data(
                         src_conn, 
@@ -143,6 +142,8 @@ def main():
                         sanitized_field=s2d.kind
                     )
 
+                # if there is a destination table/database, then create the specified destination table (replacing table if it already exists)
+                # and then update that destination table
                 else:
                     if offset == 0: 
                         create_dest_table(
