+++ code/steps/sanitize_urls/sanitization_code/pg_transfer.py
@@ -1,7 +1,6 @@
 from subprocess import PIPE,Popen
 
-# consider passing PGPASSWORD in as environment variable, along with default env from os.environ so that it doesn't error out?
-
+# creating .dmp file for specified table in specified database
 def dump_table(host_name,port,database_name,user_name,database_password,schema_name, table_name, dump_file_loc):
 
     command = f'pg_dump -d postgres://{user_name}:{database_password}@{host_name}:{port}/{database_name} -t {schema_name}.{table_name} -Fc -f {dump_file_loc}'
@@ -10,6 +9,7 @@ def dump_table(host_name,port,database_name,user_name,database_password,schema_n
 
     return p.communicate()
 
+# use .dmp file created by dump_table() to re-create table in another database
 def restore_table(host_name,port,database_name,user_name,database_password,dump_file_loc):
 
     command = f'pg_restore -d postgres://{user_name}:{database_password}@{host_name}:{port}/{database_name} {dump_file_loc}'
